+++ device-bootloaders/src/ServiceProcessorBL2/driver/i2c_controller.c
@@ -28,6 +28,7 @@
 #include "hwinc/hal_device.h"
 #include "bl2_i2c_driver.h"
 #include "interrupt.h"
+#include "delays.h"
 
 static void i2c_clock_init(ET_I2C_DEV_t *dev)
 {
@@ -171,7 +172,9 @@ int i2c_write(ET_I2C_DEV_t *dev, uint8_t regAddr, const uint8_t *txDataBuff, uin
         while (I2C_IC_STATUS_MST_ACTIVITY_GET(dev->regs->IC_STATUS) ==
                I2C_IC_STATUS_MST_ACTIVITY_MST_ACTIVITY_ACTIVE)
             ;
-
+        /* TODO: Fix I2C driver to use ack instead of adding delays after each operation 
+        delays are added to prevent SP from hanging with back to back I2C read/writes. */
+        msdelay(1);
         xSemaphoreGive(dev->bus_lock_handle);
     }
     else
@@ -219,7 +222,9 @@ int i2c_read(ET_I2C_DEV_t *dev, uint8_t regAddr, uint8_t *rxDataBuff, uint8_t rx
         {
             rxDataBuff[n] = I2C_IC_DATA_CMD_DAT_GET(dev->regs->IC_DATA_CMD);
         }
-
+        /* TODO: Fix I2C driver to use ack instead of adding delays after each operation 
+        delays are added to prevent SP from hanging with back to back I2C read/writes. */
+        msdelay(1);
         xSemaphoreGive(dev->bus_lock_handle);
     }
     else